[Fix] Align VoxCPM2 graph scheduler gate
